+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Make stream redirection optional for Invoke-CommandWithEmailWrapper
 
+### Changed
+- Make EmailFrom mandatory for Invoke-CommandWithEmailWrapper
+
 ## [1.1.0] - 2019-02-21
 ### Added
 - Unstyled table/th/td tags included in Send-HtmlMailMessage are now formatted